Weaver: netstandard -> netcore
