+++ Weaver/Xtensive.Orm.Weaver/AssemblyChecker.cs
@@ -12,7 +12,7 @@ namespace Xtensive.Orm.Weaver
   public sealed class AssemblyChecker
   {
     private readonly Dictionary<string, byte[]> frameworkAssemblies = new Dictionary<string, byte[]>(WeavingHelper.AssemblyNameComparer);
-    private readonly Dictionary<string, byte[]> netStandardAssemblies = new Dictionary<string, byte[]>(WeavingHelper.AssemblyNameComparer);
+    private readonly Dictionary<string, byte[]> netCoreAssemblies = new Dictionary<string, byte[]>(WeavingHelper.AssemblyNameComparer);
 
 
     public void RegisterFrameworkAssembly(string name, string publicKeyToken)
@@ -20,9 +20,9 @@ namespace Xtensive.Orm.Weaver
       frameworkAssemblies.Add(name, WeavingHelper.ParsePublicKeyToken(publicKeyToken));
     }
 
-    public void RegisterNetStandardAssembly(string name, string publicKeyToken)
+    public void RegisterNetCoreAssembly(string name, string publicKeyToken)
     {
-      netStandardAssemblies.Add(name, WeavingHelper.ParsePublicKeyToken(publicKeyToken));
+      netCoreAssemblies.Add(name, WeavingHelper.ParsePublicKeyToken(publicKeyToken));
     }
 
     public bool IsFrameworkAssembly(AssemblyNameReference reference)
@@ -31,10 +31,10 @@ namespace Xtensive.Orm.Weaver
       return frameworkAssemblies.TryGetValue(reference.Name, out expectedToken) && reference.HasPublicKeyToken(expectedToken);
     }
 
-    public bool IsNetStandardAssembly(AssemblyNameReference reference)
+    public bool IsNetCoreAssembly(AssemblyNameReference reference)
     {
       byte[] expectedToken;
-      return netStandardAssemblies.TryGetValue(reference.Name, out expectedToken) && reference.HasPublicKeyToken(expectedToken);
+      return netCoreAssemblies.TryGetValue(reference.Name, out expectedToken) && reference.HasPublicKeyToken(expectedToken);
     }
   }
 }
