+++ Weaver/Xtensive.Orm.Weaver/NetCoreAssemblyList.cs
@@ -1,0 +1,128 @@
+// Copyright (C) 2018-2020 Xtensive LLC.
+// This code is distributed under MIT license terms.
+// See the License.txt file in the project root for more information.
+
+namespace Xtensive.Orm.Weaver
+{
+  internal static class NetCoreAssemblyList
+  {
+    public static string[] Get()
+    {
+      return new[] {
+        "Microsoft.Win32.Primitives b03f5f7f11d50a3a",
+        "mscorlib b77a5c561934e089",
+        "netstandard cc7b13ffcd2ddd51",
+        "System.AppContext b03f5f7f11d50a3a",
+        "System.Collections.Concurrent b03f5f7f11d50a3a",
+        "System.Collections b03f5f7f11d50a3a",
+        "System.Collections.NonGeneric b03f5f7f11d50a3a",
+        "System.Collections.Specialized b03f5f7f11d50a3a",
+        "System.ComponentModel.Composition b77a5c561934e089",
+        "System.ComponentModel b03f5f7f11d50a3a",
+        "System.ComponentModel.EventBasedAsync b03f5f7f11d50a3a",
+        "System.ComponentModel.Primitives b03f5f7f11d50a3a",
+        "System.ComponentModel.TypeConverter b03f5f7f11d50a3a",
+        "System.Console b03f5f7f11d50a3a",
+        "System.Core b77a5c561934e089",
+        "System.Data.Common b03f5f7f11d50a3a",
+        "System.Data b77a5c561934e089",
+        "System.Diagnostics.Contracts b03f5f7f11d50a3a",
+        "System.Diagnostics.Debug b03f5f7f11d50a3a",
+        "System.Diagnostics.FileVersionInfo b03f5f7f11d50a3a",
+        "System.Diagnostics.Process b03f5f7f11d50a3a",
+        "System.Diagnostics.StackTrace b03f5f7f11d50a3a",
+        "System.Diagnostics.TextWriterTraceListener b03f5f7f11d50a3a",
+        "System.Diagnostics.Tools b03f5f7f11d50a3a",
+        "System.Diagnostics.TraceSource b03f5f7f11d50a3a",
+        "System.Diagnostics.Tracing b03f5f7f11d50a3a",
+        "System b77a5c561934e089",
+        "System.Drawing b03f5f7f11d50a3a",
+        "System.Drawing.Primitives b03f5f7f11d50a3a",
+        "System.Dynamic.Runtime b03f5f7f11d50a3a",
+        "System.Globalization.Calendars b03f5f7f11d50a3a",
+        "System.Globalization b03f5f7f11d50a3a",
+        "System.Globalization.Extensions b03f5f7f11d50a3a",
+        "System.IO.Compression b77a5c561934e089",
+        "System.IO.Compression.FileSystem b77a5c561934e089",
+        "System.IO.Compression.ZipFile b77a5c561934e089",
+        "System.IO b03f5f7f11d50a3a",
+        "System.IO.FileSystem b03f5f7f11d50a3a",
+        "System.IO.FileSystem.DriveInfo b03f5f7f11d50a3a",
+        "System.IO.FileSystem.Primitives b03f5f7f11d50a3a",
+        "System.IO.FileSystem.Watcher b03f5f7f11d50a3a",
+        "System.IO.IsolatedStorage b03f5f7f11d50a3a",
+        "System.IO.MemoryMappedFiles b03f5f7f11d50a3a",
+        "System.IO.Pipes b03f5f7f11d50a3a",
+        "System.IO.UnmanagedMemoryStream b03f5f7f11d50a3a",
+        "System.Linq b03f5f7f11d50a3a",
+        "System.Linq.Expressions b03f5f7f11d50a3a",
+        "System.Linq.Parallel b03f5f7f11d50a3a",
+        "System.Linq.Queryable b03f5f7f11d50a3a",
+        "System.Net b03f5f7f11d50a3a",
+        "System.Net.Http b03f5f7f11d50a3a",
+        "System.Net.NameResolution b03f5f7f11d50a3a",
+        "System.Net.NetworkInformation b03f5f7f11d50a3a",
+        "System.Net.Ping b03f5f7f11d50a3a",
+        "System.Net.Primitives b03f5f7f11d50a3a",
+        "System.Net.Requests b03f5f7f11d50a3a",
+        "System.Net.Security b03f5f7f11d50a3a",
+        "System.Net.Sockets b03f5f7f11d50a3a",
+        "System.Net.WebHeaderCollection b03f5f7f11d50a3a",
+        "System.Net.WebSockets.Client b03f5f7f11d50a3a",
+        "System.Net.WebSockets b03f5f7f11d50a3a",
+        "System.Numerics b77a5c561934e089",
+        "System.ObjectModel b03f5f7f11d50a3a",
+        "System.Reflection b03f5f7f11d50a3a",
+        "System.Reflection.Extensions b03f5f7f11d50a3a",
+        "System.Reflection.Primitives b03f5f7f11d50a3a",
+        "System.Resources.Reader b03f5f7f11d50a3a",
+        "System.Resources.ResourceManager b03f5f7f11d50a3a",
+        "System.Resources.Writer b03f5f7f11d50a3a",
+        "System.Runtime.CompilerServices.VisualC b03f5f7f11d50a3a",
+        "System.Runtime b03f5f7f11d50a3a",
+        "System.Runtime.Extensions b03f5f7f11d50a3a",
+        "System.Runtime.Handles b03f5f7f11d50a3a",
+        "System.Runtime.InteropServices b03f5f7f11d50a3a",
+        "System.Runtime.InteropServices.RuntimeInformation b03f5f7f11d50a3a",
+        "System.Runtime.Numerics b03f5f7f11d50a3a",
+        "System.Runtime.Serialization b77a5c561934e089",
+        "System.Runtime.Serialization.Formatters b03f5f7f11d50a3a",
+        "System.Runtime.Serialization.Json b03f5f7f11d50a3a",
+        "System.Runtime.Serialization.Primitives b03f5f7f11d50a3a",
+        "System.Runtime.Serialization.Xml b03f5f7f11d50a3a",
+        "System.Security.Claims b03f5f7f11d50a3a",
+        "System.Security.Cryptography.Algorithms b03f5f7f11d50a3a",
+        "System.Security.Cryptography.Csp b03f5f7f11d50a3a",
+        "System.Security.Cryptography.Encoding b03f5f7f11d50a3a",
+        "System.Security.Cryptography.Primitives b03f5f7f11d50a3a",
+        "System.Security.Cryptography.X509Certificates b03f5f7f11d50a3a",
+        "System.Security.Principal b03f5f7f11d50a3a",
+        "System.Security.SecureString b03f5f7f11d50a3a",
+        "System.ServiceModel.Web 31bf3856ad364e35",
+        "System.Text.Encoding b03f5f7f11d50a3a",
+        "System.Text.Encoding.Extensions b03f5f7f11d50a3a",
+        "System.Text.RegularExpressions b03f5f7f11d50a3a",
+        "System.Threading b03f5f7f11d50a3a",
+        "System.Threading.Overlapped b03f5f7f11d50a3a",
+        "System.Threading.Tasks b03f5f7f11d50a3a",
+        "System.Threading.Tasks.Parallel b03f5f7f11d50a3a",
+        "System.Threading.Thread b03f5f7f11d50a3a",
+        "System.Threading.ThreadPool b03f5f7f11d50a3a",
+        "System.Threading.Timer b03f5f7f11d50a3a",
+        "System.Transactions b77a5c561934e089",
+        "System.ValueTuple cc7b13ffcd2ddd51",
+        "System.Web b03f5f7f11d50a3a",
+        "System.Windows b03f5f7f11d50a3a",
+        "System.Xml b77a5c561934e089",
+        "System.Xml.Linq b77a5c561934e089",
+        "System.Xml.ReaderWriter b03f5f7f11d50a3a",
+        "System.Xml.Serialization b77a5c561934e089",
+        "System.Xml.XDocument b03f5f7f11d50a3a",
+        "System.Xml.XmlDocument b03f5f7f11d50a3a",
+        "System.Xml.XmlSerializer b03f5f7f11d50a3a",
+        "System.Xml.XPath b03f5f7f11d50a3a",
+        "System.Xml.XPath.XDocument b03f5f7f11d50a3a",
+      };
+    }
+  }
+}
